refactor: change folder disposition
