+++ src/data/interactors/user/FIndUserAdapterDb.ts
@@ -1,4 +1,4 @@
-import { FindUserRepository } from '@data/repositories/FindUserRepository';
+import { FindUserRepository } from '@data/repositories/user/FindUserRepository';
 import { FindUser } from '@domain/interactors/FindUser';
 
 export default class FindUserAdapterDb implements FindUser {
