+++ src/data/interactors/user/RemoveUserAdapterDb.ts
@@ -1,4 +1,4 @@
-import { RemoveUserRepository } from '@data/repositories/RemoveUserRepository';
+import { RemoveUserRepository } from '@data/repositories/user/RemoveUserRepository';
 import { RemoveUser } from '@domain/interactors/RemoveUser';
 
 export class RemoveUserAdapterDb implements RemoveUser {
