+++ src/data/repositories/user/CreateUserRepository.ts
@@ -1,0 +1,5 @@
+import { User } from '@domain/entities/User';
+
+export interface CreateUserRepository {
+  execute: (data: Omit<User, 'role'>) => Promise<User>
+}
